Add files via upload
This is an implementation of Kruskals Algorithm. The program reads a text file that contains the edges and weights of a graph. The Edge class will have node 1 (String), node 2 (String), and weight (Int). The text file is read line by line. For each line, I loop through the edges and create an Edge object for each edge which is then added to an array of edges. Next, the array is sorted according to the weights using Arrays.sort(). I loop through the sorted edges starting with the smallest weight and check the root of node1 and node2 using the function "findRoot". I do not print the edge if the two roots are the same, however, I print the edge making changes to the parent array and add the weight of that edge (1) to a variable that will contain the weight of the entire minimum spanning tree and then go to the next edge.  "findRoot" will check the dictionary or hashmap of the nodes and their parents.